+++ deploy/005-deploy-VTreasuryV8.ts
@@ -2,6 +2,8 @@ import { ethers } from "hardhat";
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
+import { skipSourceNetworks } from "../helpers/deploymentConfig";
+
 interface AdminAccounts {
   [key: string]: string;
 }
@@ -30,18 +32,23 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     arbitrumsepolia: "0x1426A5Ae009c4443188DA8793751024E358A61C2", // ARBITRUM_SEPOLIA MULTISIG
     arbitrumone: "0x14e0E151b33f9802b3e75b621c1457afc44DcAA0", // ARBITRUM_ONE MULTISIG
     xlayertestnet: "0x5961449d63149035aCfC0714D5155f24C9819004", // XLAYER TESTNET MULTISIG
+    zksyncsepolia: "0xa2f83de95E9F28eD443132C331B6a9C9B7a9F866", // ZKSYNC SEPOLIA MULTISIG
+    zksyncmainnet: "0x751Aa759cfBB6CE71A43b48e40e1cCcFC66Ba4aa", // ZKSYNC MAINNET MULTISIG
+    opsepolia: "0xd57365EE4E850e881229e2F8Aa405822f289e78d", // OPSEPOLIA MULTISIG
     bscmainnet: await getTimelock(),
     bsctestnet: await getTimelock(),
     hardhat: deployer,
   };
 
   const deployerSigner = await hre.ethers.getSigner(deployer);
+
   const treasuryInstance = await deploy("VTreasuryV8", {
     contract: "VTreasuryV8",
     from: deployer,
     args: [],
     log: true,
     autoMine: true,
+    skipIfAlreadyDeployed: true,
   });
 
   const adminAccount: string = acmAdminAccount[hre.network.name];
@@ -57,6 +64,6 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
 };
 
 func.tags = ["VTreasuryV8"];
-func.skip = async hre => hre.network.name === "bsctestnet" || hre.network.name === "bscmainnet";
+func.skip = skipSourceNetworks();
 
 export default func;